+++ berkeley-function-call-leaderboard/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to the Berkeley Function Calling Leaderboard will be documented in this file.
 
+
+- [Feb 9, 2025] [#901](https://github.com/ShishirPatil/gorilla/pull/901): Add new model `DeepSeek-R1` to the leaderboard.
+- [Feb 5, 2025] [#900](https://github.com/ShishirPatil/gorilla/pull/900): Add the following new models to the leaderboard:
+  - `uiuc-convai/CALM-8B`
+  - `uiuc-convai/CALM-70B`
+  - `uiuc-convai/CALM-405B`
+- [Feb 5, 2025] [#902](https://github.com/ShishirPatil/gorilla/pull/902): Add the following new models to the leaderboard:
+  - `gemini-2.0-flash-lite-preview-02-05-FC`
+  - `gemini-2.0-flash-lite-preview-02-05`
+  - `gemini-2.0-flash-001-FC`
+  - `gemini-2.0-flash-001`
+  - `gemini-2.0-pro-exp-02-05-FC`
+  - `gemini-2.0-pro-exp-02-05`
 - [Feb 2, 2025] [#898](https://github.com/ShishirPatil/gorilla/pull/898): Add the following new models to the leaderboard:
   - `o3-mini-2025-01-31-FC`
   - `o3-mini-2025-01-31`